Add files via upload
This code defines two functions, divide_image() and divide_images_in_folder(), which can be used to divide an image into four separate images and to divide all the images in a folder, respectively.

The divide_image() function takes an image file as an argument, opens the image using the Image.open() function from the PIL library, gets its dimensions using the size attribute, calculates the dimensions of the four separate images by dividing the width and height by two, and then crops the image into four separate images using the crop() function. The function saves the four separate images using the original file name plus a suffix indicating which quadrant it represents.

The divide_images_in_folder() function takes a folder path as an argument, uses the os library to get a list of image files in the folder, and then iterates through each image file, calling the divide_image() function to divide it into four separate images. The function prints out messages indicating the number of images in the folder, which image is currently being processed, and when each image has been divided.

Finally, the code calls the divide_images_in_folder() function with a sample folder path as an argument to test the function.